Clean up project structure
to prepare future integration between simulation and gui :beers:
Replaced angle(x1, y1, x2, y2) by angle(x,y)
+ minor modifications and bug fixes
